+++ dbms/src/Functions/registerFunctionsTuple.cpp
@@ -1,0 +1,15 @@
+#include <Functions/FunctionFactory.h>
+
+namespace DB
+{
+
+void registerFunctionTuple(FunctionFactory &);
+void registerFunctionTupleElement(FunctionFactory &);
+
+void registerFunctionsTuple(FunctionFactory & factory)
+{
+    registerFunctionTuple(factory);
+    registerFunctionTupleElement(factory);
+}
+
+}
